+++ app/Http/Controllers/CampaignsServicesController.php
@@ -16,7 +16,7 @@ class CampaignsServicesController extends Controller
     public function index()
     {
         // Eloquent
-        // $this->authorize('viewAny' , CampaignsServices::class);
+        $this->authorize('viewAny' , CampaignsServices::class);
         $campaignsServices = CampaignsServices::paginate(10);
         return new CampaignsServicesResourceCollection($campaignsServices);
     }
@@ -26,7 +26,7 @@ class CampaignsServicesController extends Controller
      */
     public function store(Request $request)
     {
-        // $this->authorize('create' , CampaignsServices::class);
+        $this->authorize('create' , CampaignsServices::class);
         // validate
         $valid = Validator($request->all(), [
             'amount' => 'required|numeric|min:0',
@@ -67,7 +67,7 @@ class CampaignsServicesController extends Controller
      */
     public function show(CampaignsServices $campaignsService)
     {
-        // $this->authorize('view' , $campaignsService);
+        $this->authorize('view' , $campaignsService);
         return new CampaignsServicesResource($campaignsService);
     }
 
@@ -94,7 +94,7 @@ class CampaignsServicesController extends Controller
         if (!$valid->fails()) {
             // Eloquent
             $campaignsServices = CampaignsServices::findOrFail($id);
-            // $this->authorize('update' , $campaignsServices);
+            $this->authorize('update' , $campaignsServices);
             $campaignsServices->amount = $request->input('amount');
             $campaignsServices->description = $request->input('description');
             $campaignsServices->start_date = $request->input('start_date');
@@ -119,7 +119,7 @@ class CampaignsServicesController extends Controller
     {
         // Eloquent
         $campaignsServices = CampaignsServices::findOrFail($id);
-        // $this->authorize('delete' , $campaignsServices);
+        $this->authorize('delete' , $campaignsServices);
         $deleted = $campaignsServices->delete();
         return new Response(
             [
@@ -133,7 +133,7 @@ class CampaignsServicesController extends Controller
     {
         //
         $object = CampaignsServices::onlyTrashed()->findOrFail($id);
-        // $this->authorize('restore' , $object);
+        $this->authorize('restore' , $object);
         $restored = $object->restore();
         return response()->json(['status' => $restored]);
     }
@@ -142,7 +142,7 @@ class CampaignsServicesController extends Controller
     {
         //
         $object = CampaignsServices::withTrashed()->findOrFail($id);
-        // $this->authorize('forceDelete' , $object);
+        $this->authorize('forceDelete' , $object);
         $deleted = $object->forceDelete();
         // if ($deleted && $object->image) {
         //     // Storage::delete($object->image);
@@ -154,7 +154,7 @@ class CampaignsServicesController extends Controller
 
     public function Archives()
     {
-        // $this->authorize('restore', $supplier);
+        $this->authorize('restore', $supplier);
         $cs = CampaignsServices::onlyTrashed()->get();
         $data = CampaignsServicesResource::collection($cs);
         return response()->json(['status' => true, 'message' => 'success', 'data' => $data], 200);
